old_streamlit_files/requirements-ui.txt
streamlit
pandas
numpy
aiosqlite
python-dotenv